requirements.txt
requests
tzdata
openai
langchain
langchain-community
pgvector
psycopg2-binary
tiktoken